requirements.txt
flare-capa==7.0.1
